add videoPlayer

+++ controllers/videoController.js
@@ -61,7 +61,7 @@ export const postUpload = async (req, res) => {
     description,
     creator: req.user.id,
   });
-  console.log(newVideo);
+  // console.log(newVideo);
   req.user.videos.push(newVideo.id);
   req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
@@ -74,7 +74,7 @@ export const videoDetail = async (req, res) => {
 
   try {
     const video = await Video.findById(id).populate("creator");
-    console.log(video);
+    // console.log(video);
 
     res.render("videoDetail", {
       pageTitle: video.title,
@@ -92,10 +92,14 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", {
-      pageTitle: `Edit ${video.title}`,
-      video,
-    });
+    if (video.creator.toString() !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", {
+        pageTitle: `Edit ${video.title}`,
+        video,
+      });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -128,12 +132,16 @@ export const deleteVideo = async (req, res) => {
   } = req;
 
   try {
-    await Video.findOneAndRemove({
-      _id: id,
-    });
+    const video = await Video.findById(id);
+    if (video.creator.toString() !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({
+        _id: id,
+      });
+    }
   } catch (error) {
     console.log(error);
   }
   res.redirect(routes.home);
-  // res.render("deleteVideo", { pageTitle: "Delete Video"});
 };
